+++ packages/analytics/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @happyvertical/smrt-analytics
 
+## 0.39.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @happyvertical/smrt-core@0.39.9
+  - @happyvertical/smrt-prompts@0.39.9
+  - @happyvertical/smrt-tenancy@0.39.9
+  - @happyvertical/smrt-types@0.39.9
+  - @happyvertical/smrt-ui@0.39.9
+
 ## 0.39.8
 
 ### Patch Changes