+++ packages/smrt-web/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @happyvertical/smrt-web
 
+## 0.39.9
+
 ## 0.39.8
 
 ## 0.39.7